ali-agentic-adk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
# PyPI包名
name = "ali_agentic_adk_python"
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICEN[CS]E*"]
authors = [
    { name = "AIDC", email = "mengyu.smy@alibaba-inc.com" }
]

dependencies = [
    "google-adk",
    "google-generativeai",
    "aiohttp",
    "dashscope",
    "openai",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "boto3",
    "cohere",
    "mcp>=1.15.0",
    "requests",
    "tencentcloud-sdk-python",

    # Alibaba Cloud SDKs
    "alibabacloud_ecd20200930",
    "alibabacloud_ecd20201002",
    "alibabacloud_eds_aic20230930",
    "alibabacloud_appstream_center20210218",
]

[tool.hatch.build]
exclude = ["examples/**", "tests/**", "docs/**"]

[tool.hatch.build.targets.wheel]
packages = ["src/ali_agentic_adk_python"]

[tool.hatch.build.targets.sdist]
include = [
    "src/ali_agentic_adk_python",
    "pyproject.toml",
    "README.md",
    "LICENSE",
    "NOTICE",
]
